refactor: enhance useBlockExplorerNavigation hook and streamline EmptyHistory and TxHistoryListView components
- Refactored useBlockExplorerNavigation hook to include openExplorer function for handling navigation logic based on network selection.
- Updated EmptyHistory and TxHistoryListView components to utilize the new openExplorer function, simplifying the handleOnPress logic and improving code clarity.
- Removed unnecessary imports and dependencies, enhancing overall maintainability.
